% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f2012}
\alias{f2012}
\title{\code{f2012}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38974 rows and 1883 columns.}
\source{
\url{https://meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-155}
}
\usage{
f2012
}
\description{
MEPS Household Component 2012 Full Year Consolidated Data File
}
\keyword{datasets}
