% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f2013}
\alias{f2013}
\title{\code{f2013}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36940 rows and 1790 columns.}
\source{
\url{https://meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-163}
}
\usage{
f2013
}
\description{
MEPS Household Component 2013 Full Year Consolidated Data File
}
\keyword{datasets}
