% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f2014}
\alias{f2014}
\title{\code{f2014}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34875 rows and 1838 columns.}
\source{
\url{https://meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-171}
}
\usage{
f2014
}
\description{
MEPS Household Component 2014 Full Year Consolidated Data File
}
\keyword{datasets}
